# CUDA: Faster Mixtral prompt processing