On master Mixtral prompt processing is currently always done with a batch size of 1 due to MoE. This PR makes it so that instead for batch sizes > 1 the `src1` columns are made contiguous so that the standard matrix matrix multiplication kernels can be used. This implementation is still very inefficient but significantly better than master. Partial offloading support is not implemented. ~~**Multi GPU is currently broken.**~~

Note: in the code the second dimension of `src1` is referred to as "rows". That is incorrect because with the data layout for that tensor those are actually columns. Still, I have continued the use of this terminology to keep the code consistent.

Performance:

| GPU      | Model       | Test | t/s master | t/s PR | Speedup |
|----------|-------------|------|------------|--------|---------|
| RTX 3090 | 8x7b q2_K   | pp   |        132 |    397 |    3.01 |
| RTX 3090 | 8x7b q3_K_S | pp   |         73 |    371 |    5.08 |
| P40      | 8x7b q2_K   | pp   |         38 |    115 |    3.03 |
| P40      | 8x7b q3_K_S | pp   |         19 |    108 |    5.68 |