@slaren 2023-12-19
> Note: in the code the second dimension of `src1` is referred to as "rows". That is incorrect because with the data layout for that tensor those are actually columns. Still, I have continued the use of this terminology to keep the code consistent.

I guess it depends on the perspective, but I don't really think that this is wrong. The second dimension are rows by definition. `src1` is always transposed in our matrix multiplications.