@ggerganov 2023-12-20
> The second dimension are rows by definition.

Yes, this is the convention in `ggml` - dim 0 is columns, dim 1 is rows, always. Matrix multiplication is done multiplying row-by-row, not row-by-column. It's different from mainstream conventions, but it is not wrong. I've been noticing some parts of the code that mix "rows" and "columns" in this regard - it's something that we have to eventually improve.