@JohannesGaessler 2023-12-20
I've tracked down the problem with multi GPU. It is due to peer access. Peer access is enabled/disabled by the CPU but the operations that depend on it are run asynchronously by the GPUs. Currently peer access is toggled in `ggml_cuda_op_mul_mat` based on the *current* shape of `src1`. For LLaMA-based models this is not an issue because for each number of tokens peer access has always been consistently enabled/disabled across a given eval. However, with Mixtral an expert can receive a batch size below the peer access threshold so peer access can flip during the eval which then causes a crash. I think I can do a quick fix but long term we should make it so that peer access is set per eval instead of per tensor, otherwise we would need to synchronize CPU and GPUs every time.

Regarding rows vs. columns: I would agree that if matrices are defined as transposed then the terminology is correct. I think this should be documented in the coding guidelines section of the README. My intuition is definitely different. Does the terminology only apply to the dimensions in the context of tensor or also to e.g. the CUDA code that only operates on buffers? Is there an established term for the third dimension? I have referred to it as "channel" in my code.