@JohannesGaessler 2023-12-20
My idea for a quick fix was flawed. I've `ggml_cuda_set_peer_access` and made it so that instead it is called at the beginning of each eval instead. Performance:

| GPU      | Model       | Test | t/s master | t/s PR | Speedup |
|----------|-------------|------|------------|--------|---------|
| RTX 3090 | 8x7b q2_K   | pp   |        132 |    397 |    3.01 |
| RTX 3090 | 8x7b q3_K_S | pp   |         73 |    371 |    5.08 |
| P40      | 8x7b q2_K   | pp   |         38 |    115 |    3.03 |
| P40      | 8x7b q3_K_S | pp   |         19 |    108 |    5.68 |
| 3xP40    | 8x7b q2_K   | pp   |         29 |    116 |       4 |
| 3xP40    | 8x7b q3_K_S | pp   |         23 |    111 |    4.83 |
| 3xP40    | 8x7b q8_0   | pp   |         24 |    126 |    5.25 |