@slaren 2023-12-20
> I think I can do a quick fix but long term we should make it so that peer access is set per eval instead of per tensor, otherwise we would need to synchronize CPU and GPUs every time.

So if I understand correctly, the issue is that changing peer access while there are operations queued in a stream causes CUDA to crash. If so, have you tried if adding a `cudaDeviceSynchronize` in `ggml_cuda_set_peer_access` when `peer_access_enabled` changes affects performance significantly?