@JohannesGaessler 2023-12-20
I did a quick test and synchronizing the GPUs every time is barely faster than master:

| GPU   | Model       | Test | t/s PR | t/s synchronize | Speedup |
|-------|-------------|------|--------|-----------------|---------|
| 3xP40 | 8x7b q2_K   | pp   |    116 |              26 |    0.22 |
| 3xP40 | 8x7b q3_K_S | pp   |    111 |              31 |    0.28 |
| 3xP40 | 8x7b q8_0   | pp   |    126 |              34 |    0.27 |

I don't know what percentage of the overhead is from synchronization vs. peer access toggling.
But if the overhead from peer access toggling is high then we probably wouldn't want to do this on a per-tensor basis anyways because the overhead would be more than the time saved on memcpys.