@slaren 2023-12-20
Still, it would be better to handle this inside ggml-cuda. We should keep backend-specific code to a minimum in llama.cpp. We could also add a flag to `ggml_cuda_mul_mat` to avoid changing peer access when called from `ggml_cuda_mul_mat_id`, or maybe just setting the peer access in `ggml_cuda_compute_forward`. Eventually we will change ggml-cuda to process entire graphs at a time with ggml-backend, and then we can just do it once at the start of the graph.