@JohannesGaessler 2023-12-20
Good points. I have pushed a version with synchronization but moved the call to `ggml_cuda_set_peer_access` to `ggml_cuda_compute_forward`. The performance is the same as with the previous version of the PR. I was concerned about the automatism severely gimping performance for some sequences of matrix multiplications but it seems there is a long term solution.