@slaren 2023-12-20
I see a ~5.8x speedup with 3090 Ti under WSL2 (tested with `perplexity`).