@ggerganov 2023-12-21
> Regarding rows vs. columns: I would agree that if matrices are defined as transposed then the terminology is correct. I think this should be documented in the coding guidelines section of the README. My intuition is definitely different. Does the terminology only apply to the dimensions in the context of tensor or also to e.g. the CUDA code that only operates on buffers? Is there an established term for the third dimension? I have referred to it as "channel" in my code.

Updated the coding guidelines as suggested. For the third dimension we tend to use "matrices" in `ggml.c`, but I guess "channels" is fine too.